scripts/tiktok/requirements.txt
kokoro-onnx==0.3.7
soundfile==0.12.1
moviepy==2.1.2
numpy==2.1.3
pillow==11.0.0
openai-whisper==20240930
requests==2.32.3
